sv_rounds: Set round status after setting up players
The player setup routine includes killing and respawning
all players, but at that time, the round status has been
set to "RUNNING" already.

That behaviour might cause problems for plugins that want
to do stuff on player death, but only if the round is in
progress (which it isn't yet).

Change the order of commands so that the round is only
set to run after all the setup has been done.